+++ core/factgen.h
@@ -224,6 +224,6 @@ struct factCheckLengthPNCall
    struct expr               *FactGenGetfield(Environment *,struct lhsParseNode *);
    struct expr               *FactGenGetvar(Environment *,struct lhsParseNode *,int);
    struct expr               *FactGenCheckLength(Environment *,struct lhsParseNode *);
-   struct expr               *FactGenCheckZeroLength(Environment *,unsigned);
+   struct expr               *FactGenCheckZeroLength(Environment *,unsigned short);
 
 #endif /* _H_factgen */
