+++ core/factqpsr.c
@@ -404,7 +404,7 @@ static bool ReplaceTemplateNameWithReference(
   {
    const char *theTemplateName;
    void *theDeftemplate;
-   int count;
+   unsigned int count;
 
    if (theExp->type == SYMBOL_TYPE)
      {
@@ -633,8 +633,8 @@ static void ReplaceFactVariables(
            {
             bexp->type = FCALL;
             bexp->value = rindx_func;
-            eptr = GenConstant(theEnv,INTEGER_TYPE,CreateInteger(theEnv,(long long) ndepth));
-            eptr->nextArg = GenConstant(theEnv,INTEGER_TYPE,CreateInteger(theEnv,(long long) posn));
+            eptr = GenConstant(theEnv,INTEGER_TYPE,CreateInteger(theEnv,ndepth));
+            eptr->nextArg = GenConstant(theEnv,INTEGER_TYPE,CreateInteger(theEnv,posn));
             bexp->argList = eptr;
            }
          else if (sdirect == true)
@@ -708,9 +708,9 @@ static void ReplaceSlotReference(
             CloseStringSource(theEnv,"query-var");
             theExp->type = FCALL;
             theExp->value = func;
-            theExp->argList = GenConstant(theEnv,INTEGER_TYPE,CreateInteger(theEnv,(long long) ndepth));
+            theExp->argList = GenConstant(theEnv,INTEGER_TYPE,CreateInteger(theEnv,ndepth));
             theExp->argList->nextArg =
-              GenConstant(theEnv,INTEGER_TYPE,CreateInteger(theEnv,(long long) posn));
+              GenConstant(theEnv,INTEGER_TYPE,CreateInteger(theEnv,posn));
             theExp->argList->nextArg->nextArg = GenConstant(theEnv,TokenTypeToType(itkn.tknType),itkn.value);
             break;
            }
