+++ core/msgfun.h
@@ -51,8 +51,8 @@
 
 typedef struct handlerSlotReference
   {
-   long classID;
-   long slotID;
+   unsigned classID;
+   unsigned short slotID;
   } HANDLER_SLOT_REFERENCE;
 
 #include "msgpass.h"
@@ -80,9 +80,9 @@ typedef struct handlerSlotReference
    void             SlotVisibilityViolationError(Environment *,SlotDescriptor *,Defclass *,bool);
 
 #if ! RUN_TIME
-   void             NewSystemHandler(Environment *,const char *,const char *,const char *,int);
+   void             NewSystemHandler(Environment *,const char *,const char *,const char *,unsigned short);
    DefmessageHandler
-                   *InsertHandlerHeader(Environment *,Defclass *,CLIPSLexeme *,int);
+                   *InsertHandlerHeader(Environment *,Defclass *,CLIPSLexeme *,unsigned);
 #endif
 
 #if (! BLOAD_ONLY) && (! RUN_TIME)
@@ -92,7 +92,7 @@ typedef struct handlerSlotReference
    bool             DeleteHandler(Environment *,Defclass *,CLIPSLexeme *,int,bool);
    void             DeallocateMarkedHandlers(Environment *,Defclass *);
 #endif
-   unsigned         HandlerType(Environment *,const char *,const char *);
+   unsigned short   HandlerType(Environment *,const char *,const char *);
    bool             CheckCurrentMessage(Environment *,const char *,bool);
    void             PrintHandler(Environment *,const char *,DefmessageHandler *,bool);
    DefmessageHandler
